requirements.txt
mmcv_full==1.7.2
mmdet==2.28.2
numpy==1.23.0
opencv_python==4.9.0.80
pyquaternion==0.9.9
scipy==1.10.1
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tqdm
